feat:商家端-商品库存添加接口，下架可减加库存，商家只能加库存

+++ business/member-api/src/main/java/com/lawu/eshop/member/api/controller/ProductController.java
@@ -98,7 +98,7 @@ public class ProductController extends BaseController {
 				//Result<com.lawu.eshop.product.dto.CommentProductInfoDTO> product = productService.selectCommentProductInfo(comment.getProductModelId());
 				boolean flag = true;
 				for(MemberProductModelDTO model : spec){
-					if(comment.getProductModelId() == model.getId()){
+					if(comment.getProductModelId().longValue() == model.getId().longValue()){
 						cidto.setProductModel(model.getName());
 						flag = false;
 						break;
